converted project to use a service that is depdency injected
