+++ src/app/investment-results/investment-results.component.html
@@ -13,11 +13,11 @@
         @for (result of investmentResultsData; track result.year) {
           <tr>
             <td>{{ result.year }}</td>
-            <td>{{ result.interest }}</td>
-            <td>{{ result.valueEndOfYear }}</td>
-            <td>{{ result.annualInvestment }}</td>
-            <td>{{ result.totalInterest }}</td>
-            <td>{{ result.totalAmountInvested }}</td>
+            <td>{{ result.interest | currency }}</td>
+            <td>{{ result.valueEndOfYear | currency }}</td>
+            <td>{{ result.annualInvestment | currency }}</td>
+            <td>{{ result.totalInterest | currency }}</td>
+            <td>{{ result.totalAmountInvested | currency }}</td>
           </tr>
         }
       </tbody>
